Simplify fallback behavior option & errors displayed in UI
Moved away from displaying errors in obtrusive modal alerts and instead primarily use toast notifications. Also consolidated options for the “config not found behavior” setting as a result.

The UI for toast-style notifications in Nova was changed recently (either verision 10 or 11, I can’t remember) and is harder to miss now. But mostly the modal alert was just too annoying and has no way to avoid repeating the same message.